QA_Generator/requirements.txt
streamlit
requests
pandas
PyPDF2
git+https://github.com/tomquirk/linkedin-api.git
reportlab
python-docx

openai==0.27.2
SQLAlchemy==2.0.9
stack-data==0.6.2
stripe==5.4.0
urllib3
virtualenv
websocket-client
Werkzeug
WTForms
load_dotenv
google-cloud-datastore==2.7.1
google-auth==2.17.2
requests==2.28.2
google_auth_oauthlib
google_auth_httplib2
google-api-python-client
paypalrestsdk